Move DeviceLists type to synapse.types
So that we can use it elsewhere.
