+++ synapse/types.py
@@ -19,6 +19,7 @@ from typing import (
     TYPE_CHECKING,
     Any,
     ClassVar,
+    Collection,
     Dict,
     List,
     Mapping,
@@ -741,6 +742,21 @@ class ReadReceipt:
     user_id: str
     event_ids: List[str]
     data: JsonDict
+
+
+@attr.s(slots=True, frozen=True, auto_attribs=True)
+class DeviceLists:
+    """
+    Attributes:
+        changed: List of user_ids whose devices may have changed
+        left: List of user_ids whose devices we no longer track
+    """
+
+    changed: Collection[str]
+    left: Collection[str]
+
+    def __bool__(self) -> bool:
+        return bool(self.changed or self.left)
 
 
 def get_verify_key_from_cross_signing_key(key_info):
